forensic-image-detection/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
Werkzeug==2.3.8 # Pastikan kompatibel dengan Flask Anda
Jinja2==3.1.2
SQLAlchemy==2.0.23
WTForms==3.1.2
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0
eventlet==0.33.3 # Atau gevent, untuk Flask-SocketIO di production
numpy==1.26.2
opencv-python==4.8.1.78 # Atau opencv-python-headless
Pillow==10.1.0
scikit-image==0.22.0
scipy==1.11.4
exifread==3.0.0
matplotlib==3.8.2
python-docx==1.1.0
# docx2pdf # Jika ingin konversi PDF dari DOCX, opsional, install manual jika perlu
# pywin32 # Jika di Windows dan ingin konversi PDF via MS Word, opsional
# pandas # Jika ada bagian sistem deteksi yang butuh
scikit-learn==1.3.2 # Untuk classification.py dan copy_move_detection.py